[WIP] move data fetching to parent component
