+++ package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@material-ui/core": "^4.11.3",
     "algosdk": "^1.8.1",
+    "axios": "^0.21.1",
     "gatsby": "^2.26.1",
     "gatsby-plugin-prettier-eslint": "^1.0.4",
     "prettier": "^2.2.1",
