+++ src/components/ASAList/ASAList.component.jsx
@@ -1,19 +1,8 @@
-import React, { useState, useContext, useEffect } from 'react';
+import React from 'react';
 import { useTable } from 'react-table';
-import axios from 'axios';
-import AlgoSignerContext from '../../contexts/algosigner.context';
-import ASATransactions from '../ASATransactions/ASATransactions.component';
 
-import AlgoSdk from '../../services/AlgoSdk';
-
-const ASAList = () => {
-  const [accountDetails, setAccountDetails] = useState({});
-  const [createdAssets, setCreatedAssets] = useState([]);
-  const ctx = useContext(AlgoSignerContext);
-  const [loading, setLoading] = useState(false);
-
-  const data = React.useMemo(() => createdAssets, [createdAssets]);
-
+const ASAList = ({ data }) => {
+  console.log('// data ', data);
   const columns = React.useMemo(
     () => [
       {
@@ -22,32 +11,32 @@ const ASAList = () => {
       },
       {
         Header: 'Name',
-        accessor: 'params.name', // accessor is the "key" in the data
-      },
-      {
-        Header: 'Total Balance',
         accessor: 'amount',
       },
-      {
-        Header: 'Decimals',
-        accessor: 'params.decimals',
-      },
-      {
-        Header: 'Whitelisted addresses',
-        accessor: 'whitelistedAddresses',
-      },
-      {
-        Header: 'Addresses With Balance',
-        accessor: 'addressedWithBalance',
-      },
-      {
-        Header: 'Assets Frozen',
-        accessor: 'assetsFrozen',
-      },
-      {
-        Header: 'Assets Revoked',
-        accessor: 'assetsRevoked',
-      },
+      // {
+      //   Header: 'Total Balance',
+      //   accessor: 'amount',
+      // },
+      // {
+      //   Header: 'Decimals',
+      //   accessor: 'params.decimals',
+      // },
+      // {
+      //   Header: 'Whitelisted addresses',
+      //   accessor: 'whitelistedAddresses',
+      // },
+      // {
+      //   Header: 'Addresses With Balance',
+      //   accessor: 'addressedWithBalance',
+      // },
+      // {
+      //   Header: 'Assets Frozen',
+      //   accessor: 'assetsFrozen',
+      // },
+      // {
+      //   Header: 'Assets Revoked',
+      //   accessor: 'assetsRevoked',
+      // },
     ],
     [],
   );
@@ -60,76 +49,6 @@ const ASAList = () => {
     prepareRow,
   } = useTable({ columns, data });
 
-  useEffect(() => {
-    console.log('ASAList address param ', ctx.currentAddress);
-
-    axios
-      .get(
-        `https://api.testnet.algoexplorer.io/idx2/v2/accounts/${ctx.currentAddress}`,
-      )
-      .then((res) => {
-        console.log(res.data.account);
-        setAccountDetails(res.data.account);
-        setCreatedAssets(res.data.account.assets);
-      })
-      .catch((e) => {
-        console.error(e);
-      });
-
-    if (ctx.currentAddress) {
-      AlgoSdk.getAccountDetailsIndexer(ctx.currentAddress)
-        .then((ad) => {
-          console.log('///', ad);
-          setAccountDetails(ad);
-        })
-        .catch((e) => {
-          console.error(e);
-        });
-
-      // AlgoSdk.getAccountInformation(ctx.currentAddress).then(
-      // 	(accountDetails) => {
-      // 		console.log(accountDetails);
-      // 		setAccountDetails(accountDetails);
-      // 		const columns = Object.keys(accountDetails);
-      // 		console.log(columns);
-      // 		accountDetails['created-assets'].map((item) => {
-      // 			console.log(item.index);
-      // 			console.log(item.params.name, item.params.total);
-
-      // 			axios
-      // 				.get(
-      // 					'https://api.testnet.algoexplorer.io/idx2/v2/assets/13672793/balances'
-      // 				)
-      // 				.then((res) => {
-      // 					console.log(res);
-      // 					const newBalances = res.data.balances.map((obj) => obj.address);
-      // 					console.log(newBalances);
-      // 					setAccountDetails({ balances: newBalances });
-      // 				});
-      // 		});
-      // 	}
-      // );
-
-      // async function fetchASACreated() {
-      // 	await AlgoSigner.connect();
-
-      // 	await AlgoSigner.accounts({
-      // 		ledger: ledger,
-      // 	})
-      // 		.then((AlgoSignerWallet) => {
-      // 			console.log(JSON.stringify(AlgoSignerWallet));
-      // 			setWallet(AlgoSignerWallet);
-      // 			console.log(JSON.stringify(wallet));
-      // 		})
-      // 		.catch((e) => {
-      // 			console.error(e);
-      // 		});
-      // }
-
-      // fetchWallet();
-    }
-  }, [ctx.currentAddress]);
-
   return (
     <table {...getTableProps()} style={{ border: 'solid 1px blue' }}>
       <thead>
@@ -139,10 +58,10 @@ const ASAList = () => {
               <th
                 {...column.getHeaderProps()}
                 style={{
-								  borderBottom: 'solid 3px red',
-								  background: 'aliceblue',
-								  color: 'black',
-								  fontWeight: 'bold',
+                  borderBottom: 'solid 3px red',
+                  background: 'aliceblue',
+                  color: 'black',
+                  fontWeight: 'bold',
                 }}
               >
                 {column.render('Header')}
@@ -153,27 +72,26 @@ const ASAList = () => {
       </thead>
       <tbody {...getTableBodyProps()}>
         {rows.map((row) => {
-				  prepareRow(row);
-				  return (
-  <tr {...row.getRowProps()}>
-    {row.cells.map((cell) => (
-      <td
-        {...cell.getCellProps()}
-        style={{
-									  padding: '10px',
-									  border: 'solid 1px gray',
-									  background: 'papayawhip',
-        }}
-      >
-        {cell.render('Cell')}
-      </td>
-    ))}
-  </tr>
-				  );
+          prepareRow(row);
+          return (
+            <tr {...row.getRowProps()}>
+              {row.cells.map((cell) => (
+                <td
+                  {...cell.getCellProps()}
+                  style={{
+                    padding: '10px',
+                    border: 'solid 1px gray',
+                    background: 'papayawhip',
+                  }}
+                >
+                  {cell.render('Cell')}
+                </td>
+              ))}
+            </tr>
+          );
         })}
       </tbody>
     </table>
   );
 };
-
 export default ASAList;
