+++ src/components/ASATransactions/ASATransactions.component.jsx
@@ -8,21 +8,7 @@ import { AccountDetailsContext } from '../../contexts/accountDetails.context';
 
 import AlgoSdk from '../../services/AlgoSdk';
 
-const ASATransactions = (asaId = []) => {
-  const ctx = useContext(AlgoSignerContext);
-  const [loading, setLoading] = useState(false);
-  const [transactions, setTransactions] = useState([]);
-
-  const getASATransactions = async () => {
-    const res = await axios.get(
-      'https://api.testnet.algoexplorer.io/idx2/v2/assets/13672793/transactions',
-    );
-    console.log(`getASATransactions ${res.data.transactions}`);
-    setTransactions(res.data.transactions);
-  };
-
-  const data = React.useMemo(() => transactions, [transactions]);
-
+const ASATransactions = (data) => {
   const columns = React.useMemo(
     () => [
       {
@@ -65,28 +51,6 @@ const ASATransactions = (asaId = []) => {
     [],
   );
 
-  useEffect(() => {
-    axios
-      .get(
-        'https://api.testnet.algoexplorer.io/idx2/v2/assets/13672793/transactions',
-      )
-      .then((res) => {
-        console.log(res);
-        console.log(
-          `ASATransactions for asset 13672793 ${JSON.stringify(
-            res.data.transactions,
-          )}`,
-        );
-        setTransactions(res.data.transactions);
-        // setAccountDetails(res.data.account['created-assets']);
-      })
-      .catch((e) => {
-        console.error(e);
-      });
-
-    // getASATransactions();
-  }, []);
-
   const {
     // getTableProps,
     // getTableBodyProps,
@@ -116,12 +80,12 @@ const ASATransactions = (asaId = []) => {
               <th
                 {...column.getHeaderProps()}
                 style={{
-								  borderBottom: 'solid 3px red',
-								  background: 'aliceblue',
-								  color: 'black',
-								  fontWeight: 'bold',
-								  width: '15px',
-								  maxWidth: '30px',
+                  borderBottom: 'solid 3px red',
+                  background: 'aliceblue',
+                  color: 'black',
+                  fontWeight: 'bold',
+                  width: '15px',
+                  maxWidth: '30px',
                 }}
               >
                 {column.render('Header')}
@@ -132,24 +96,24 @@ const ASATransactions = (asaId = []) => {
       </thead>
       <tbody {...getTableBodyProps()}>
         {rows.map((row) => {
-				  prepareRow(row);
-				  return (
-  <tr {...row.getRowProps()}>
-    {row.cells.map((cell) => (
-      <td
-        {...cell.getCellProps()}
-        style={{
-									  padding: '10px',
-									  border: 'solid 1px gray',
-									  background: 'papayawhip',
-									  width: '15px',
-        }}
-      >
-        {cell.render('Cell')}
-      </td>
-    ))}
-  </tr>
-				  );
+          prepareRow(row);
+          return (
+            <tr {...row.getRowProps()}>
+              {row.cells.map((cell) => (
+                <td
+                  {...cell.getCellProps()}
+                  style={{
+                    padding: '10px',
+                    border: 'solid 1px gray',
+                    background: 'papayawhip',
+                    width: '15px',
+                  }}
+                >
+                  {cell.render('Cell')}
+                </td>
+              ))}
+            </tr>
+          );
         })}
       </tbody>
     </table>
